pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "cvae-mnist"
version = "0.0.1"
description = "The project aims to explore the potential of Conditional Variational Autoencoders (C-VAEs) for generating synthetic data from specific image classes in the MNIST dataset."
authors = [
  { name = "moreira-and" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
dependencies = [
    "loguru",
    "pip",
    "python-dotenv",
    "tqdm",
    "typer",
    "torch (>=2.8.0,<3.0.0)",
    "torchvision (>=0.23.0,<0.24.0)",
    "numpy (>=2.0,<2.3)",
    "pandas (>=2.3.2,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "matplotlib (>=3.10.6,<4.0.0)",
    "mlflow (>=3.3.2,<4.0.0)",
]
requires-python = "~=3.10.0"

[tool.poetry]
packages = [
  { include = "cvae", from = "src" }
]

[tool.poetry.group.dev.dependencies]
black = "*"
flake8 = "*"
pytest = "*"
isort = "*"

[project.scripts]
dataset = "cvae.cli:dataset_main"
train = "cvae.cli:train_main"
plot = "cvae.cli:plot_main"
gen   = "cvae.cli:gen_main"
experiment = "cvae.cli:experiment_main"


[tool.black]
line-length = 99
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
)/
'''

[tool.isort]
profile = "black"
known_first_party = ["src"]
force_sort_within_sections = true


